optimized hero section
